+++ src/main/java/net/kyori/registry/api/WithDefaultIdentifier.java
@@ -21,28 +21,21 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry;
+package net.kyori.registry.api;
 
-import net.kyori.registry.api.Registry;
-import net.kyori.registry.impl.DefaultableRegistry;
-import net.kyori.registry.impl.RegistryImpl;
-import org.junit.jupiter.api.Test;
+import org.checkerframework.checker.nullness.qual.NonNull;
 
-import static org.junit.jupiter.api.Assertions.*;
-
-class DefaultableRegistryTest {
-  @Test
-  void testDefaultKeyValue() {
-    final DefaultableRegistry<String, String> container = new DefaultableRegistry<>(Registry.create(), "_default");
-    assertEquals("_default", container.defaultKey());
-
-    assertNull(container.get("aaa"));
-    assertNull(container.get("ccc"));
-
-    container.register("_default", "bbb");
-    assertEquals("bbb", container.get("aaa"));
-    assertNotNull(container.get("ccc"));
-
-    assertEquals("bbb", container.getOrDefault("ccc"));
-  }
+/**
+ * A read-only component of an id registry with a default key and value.
+ *
+ * @param <V> the value type
+ */
+public interface WithDefaultIdentifier<V> {
+    /**
+     * Gets the id for {@code value}.
+     *
+     * @param value the value
+     * @return the id
+     */
+    int idOrDefault(final @NonNull V value);
 }
