+++ src/main/java/net/kyori/registry/api/WithDefaultValue.java
@@ -1,8 +1,9 @@
-package net.kyori.registry.api.defaultable;
+package net.kyori.registry.api;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 
-public interface Defaultable<K, V> {
+public interface WithDefaultValue<K, V> {
     @NonNull K defaultKey();
+
     @NonNull V getOrDefault(@NonNull K key);
 }
