+++ src/test/java/net/kyori/registry/id/map/IdMapImplTest.java
@@ -28,29 +28,27 @@ import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
 import net.kyori.registry.api.map.IdMap;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.*;
 
 class IdMapImplTest {
-  private static final int DEFAULT = -1000;
-  private final IdMap<String> map = IdMap.create(new Int2ObjectOpenHashMap<>(), new Object2IntOpenHashMap<String>() {
-    {
-      this.defaultReturnValue(DEFAULT);
+    private static final int DEFAULT = -1000;
+    private final IdMap<String> map = IdMap.create(new Int2ObjectOpenHashMap<>(), new Object2IntOpenHashMap<String>() {
+        {
+            this.defaultReturnValue(DEFAULT);
+        }
+    }, value -> value == -1000);
+
+    @Test
+    void testId() {
+        assertFalse(this.map.id("foo").isPresent());
+        this.map.put(32, "foo");
+        assertEquals(32, this.map.id("foo").orElse(DEFAULT));
     }
-  }, value -> value == -1000);
 
-  @Test
-  void testId() {
-    assertFalse(this.map.id("foo").isPresent());
-    this.map.put(32, "foo");
-    assertEquals(32, this.map.id("foo").orElse(DEFAULT));
-  }
-
-  @Test
-  void testGet() {
-    assertNull(this.map.get(32));
-    this.map.put(32, "foo");
-    assertEquals("foo", this.map.get(32));
-  }
+    @Test
+    void testGet() {
+        assertNull(this.map.get(32));
+        this.map.put(32, "foo");
+        assertEquals("foo", this.map.get(32));
+    }
 }
